fix laravel

+++ Laravel/database/migrations/2023_01_13_134955_create_pizzas_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('pizzasoort');
             $table->string('toppings');
-            $table->integer('prijs');
+            $table->decimal('prijs');
         });
     }
